Rename control sys AH namespace and library to Measurements
