+++ src/ControlSystem/ControlErrors/Shape.hpp
@@ -8,7 +8,7 @@
 #include <pup.h>
 #include <string>
 
-#include "ControlSystem/ApparentHorizons/BothHorizons.hpp"
+#include "ControlSystem/Measurements/BothHorizons.hpp"
 #include "ControlSystem/Protocols/ControlError.hpp"
 #include "ControlSystem/Tags.hpp"
 #include "DataStructures/DataVector.hpp"
