+++ src/ControlSystem/Measurements/CMakeLists.txt
@@ -1,7 +1,7 @@
 # Distributed under the MIT License.
 # See LICENSE.txt for details.
 
-set(LIBRARY ControlSystemApparentHorizons)
+set(LIBRARY ControlSystemMeasurements)
 
 add_spectre_library(${LIBRARY} INTERFACE)
