+++ src/ControlSystem/Systems/Expansion.hpp
@@ -9,10 +9,10 @@
 #include <string>
 
 #include "ApparentHorizons/Tags.hpp"
-#include "ControlSystem/ApparentHorizons/BothHorizons.hpp"
 #include "ControlSystem/Component.hpp"
 #include "ControlSystem/ControlErrors/Expansion.hpp"
 #include "ControlSystem/DataVectorHelpers.hpp"
+#include "ControlSystem/Measurements/BothHorizons.hpp"
 #include "ControlSystem/Protocols/ControlError.hpp"
 #include "ControlSystem/Protocols/ControlSystem.hpp"
 #include "ControlSystem/Protocols/Measurement.hpp"
@@ -53,7 +53,8 @@ namespace control_system::Systems {
  *   simulation
  * - Currently both these objects must be black holes
  * - Currently this control system can only be used with the \link
- *   control_system::ah::BothHorizons BothHorizons \endlink measurement
+ *   control_system::measurements::BothHorizons BothHorizons \endlink
+ * measurement
  * - Currently this control system can only be used with the \link
  *   control_system::ControlErrors::Expansion Expansion \endlink control error
  */
@@ -74,7 +75,7 @@ struct Expansion : tt::ConformsTo<protocols::ControlSystem> {
     return name();
   }
 
-  using measurement = ah::BothHorizons;
+  using measurement = measurements::BothHorizons;
   static_assert(
       tt::conforms_to_v<measurement, control_system::protocols::Measurement>);
 
@@ -97,7 +98,7 @@ struct Expansion : tt::ConformsTo<protocols::ControlSystem> {
         tmpl::list<StrahlkorperTags::Strahlkorper<Frame::Grid>>;
 
     template <::domain::ObjectLabel Horizon, typename Metavariables>
-    static void apply(ah::BothHorizons::FindHorizon<Horizon> /*meta*/,
+    static void apply(measurements::BothHorizons::FindHorizon<Horizon> /*meta*/,
                       const Strahlkorper<Frame::Grid>& horizon_strahlkorper,
                       Parallel::GlobalCache<Metavariables>& cache,
                       const LinkedMessageId<double>& measurement_id) {
