+++ src/ControlSystem/Systems/Rotation.hpp
@@ -9,10 +9,10 @@
 #include <string>
 
 #include "ApparentHorizons/Tags.hpp"
-#include "ControlSystem/ApparentHorizons/BothHorizons.hpp"
 #include "ControlSystem/Component.hpp"
 #include "ControlSystem/ControlErrors/Rotation.hpp"
 #include "ControlSystem/DataVectorHelpers.hpp"
+#include "ControlSystem/Measurements/BothHorizons.hpp"
 #include "ControlSystem/Protocols/ControlError.hpp"
 #include "ControlSystem/Protocols/ControlSystem.hpp"
 #include "ControlSystem/Protocols/Measurement.hpp"
@@ -52,7 +52,8 @@ namespace control_system::Systems {
  *   simulation
  * - Currently both these objects must be black holes
  * - Currently this control system can only be used with the \link
- *   control_system::ah::BothHorizons BothHorizons \endlink measurement
+ *   control_system::measurements::BothHorizons BothHorizons \endlink
+ * measurement
  * - Currently this control system can only be used with the \link
  *   control_system::ControlErrors::Rotation Rotation \endlink control error
  *
@@ -80,7 +81,7 @@ struct Rotation : tt::ConformsTo<protocols::ControlSystem> {
     return component == 0 ? "x" : component == 1 ? "y" : "z";
   }
 
-  using measurement = ah::BothHorizons;
+  using measurement = measurements::BothHorizons;
   static_assert(
       tt::conforms_to_v<measurement, control_system::protocols::Measurement>);
 
@@ -103,7 +104,7 @@ struct Rotation : tt::ConformsTo<protocols::ControlSystem> {
         tmpl::list<StrahlkorperTags::Strahlkorper<Frame::Grid>>;
 
     template <::domain::ObjectLabel Horizon, typename Metavariables>
-    static void apply(ah::BothHorizons::FindHorizon<Horizon> /*meta*/,
+    static void apply(measurements::BothHorizons::FindHorizon<Horizon> /*meta*/,
                       const Strahlkorper<Frame::Grid>& strahlkorper,
                       Parallel::GlobalCache<Metavariables>& cache,
                       const LinkedMessageId<double>& measurement_id) {
