+++ tests/Unit/ControlSystem/Test_Measurements.cpp
@@ -3,7 +3,7 @@
 
 #include "Framework/TestingFramework.hpp"
 
-#include "ControlSystem/ApparentHorizons/BothHorizons.hpp"
+#include "ControlSystem/Measurements/BothHorizons.hpp"
 #include "Domain/ObjectLabel.hpp"
 #include "Helpers/ControlSystem/Examples.hpp"
 #include "ParallelAlgorithms/Interpolation/Protocols/InterpolationTargetTag.hpp"
@@ -11,7 +11,7 @@
 
 static_assert(
     tt::assert_conforms_to_v<
-        control_system::ah::BothHorizons::
+        control_system::measurements::BothHorizons::
             FindHorizon<::domain::ObjectLabel::A>::interpolation_target_tag<
                 tmpl::list<control_system::TestHelpers::ExampleControlSystem>>,
         intrp::protocols::InterpolationTargetTag>);
